requirements.txt
diff-match-patch==20121119
Django==1.10.0
django-annoying==0.10.3
django-bower==5.2.0
django-calendarium==1.3.4
django-filer==1.2.7
django-filter==1.0.4
django-import-export==0.5.1
django-jet==1.0.6
django-libs==1.67.12
django-mptt==0.8.7
django-polymorphic==1.0.2
django-scheduler==0.8.5
easy-thumbnails==2.3
et-xmlfile==1.0.1
icalendar==3.11.4
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
Pillow==4.1.1
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
six==1.10.0
tablib==0.11.5
unicodecsv==0.14.1
Unidecode==0.4.21
xlrd==1.0.0
xlwt==1.2.0
